CairoBasic sample working on OS X again

+++ samples/CairoBasic/src/CairoBasicApp.cpp
@@ -22,8 +22,8 @@ class Flower {
 		for( int petal = 0; petal < mNumPetals; ++petal ) {
 			ctx.newSubPath();
 			float petalAngle = ( petal / (float)mNumPetals ) * 2 * M_PI;
-			vec2 outsideCircleCenter = mLoc + vec2( 1, 0 ) * cos( petalAngle ) * mRadius + vec2( 0, 1 ) * sin( petalAngle ) * mRadius;
-			vec2 insideCircleCenter = mLoc + vec2( 1, 0 ) * cos( petalAngle ) * mPetalInsideRadius + vec2( 0, 1 ) * sin( petalAngle ) * mPetalInsideRadius;
+			vec2 outsideCircleCenter = mLoc + vec2( 1, 0 ) * (float)cos( petalAngle ) * mRadius + vec2( 0, 1 ) * (float)sin( petalAngle ) * mRadius;
+			vec2 insideCircleCenter = mLoc + vec2( 1, 0 ) * (float)cos( petalAngle ) * mPetalInsideRadius + vec2( 0, 1 ) * (float)sin( petalAngle ) * mPetalInsideRadius;
 			ctx.arc( outsideCircleCenter, mPetalOutsideRadius, petalAngle + M_PI / 2 + M_PI, petalAngle + M_PI / 2 );
 			ctx.arc( insideCircleCenter, mPetalInsideRadius, petalAngle + M_PI / 2, petalAngle + M_PI / 2 + M_PI );
 			ctx.closePath();
